endpoints/share/useShareToken/requirements.txt
requests
boto3
PyJWT